Support multiple step functions and agent types
The semantics of multiple functions are currently unclear. In
particular it's not clear when the "in" and "out" buffers should
be swapped. In some cases this should happen only after all step
functions have run, but something it should happen between them.
Maybe this needs some kind of explicit annotation?