Add incomplete Kerationcyte model
The model is not completely ported and we're still missing some
necessary language features. I've already added support for global
arrays today, which was one of the large missing pieces. Another
thing that's required is the ability to add an agent inside the
step function. Furthermore, at least to support the way the FlameGPU
model is currently written, we might want to introduce support for
structs that are not agents.