Issue #2: Introduce environment min+max
If "min" is not specified, it's assumed to be zero.

This is only the syntactic support, it's not yet properly used
in the backends. Some places where this is needed, such as the
FlameGPU model file, will be somewhat tricky to handle, as we
need a constexpr evaluator for that.