Support float + double in supporting backends
By default all backends now consistently use double numbers for
computation. The configuration option "-C use_float=true" can be
used to use floats instead. This configuration option is supported
by the c, flame and flamegpu backends, but not the mason and dmason
backends. This is also why we use double as the default, it's not
a good choice, but it's the only one supported by all backends.