Add a dummy game of life implementation
We only support a continuous environment, so have to do this using
floating-point positions.

This will need some other kind of visualization than the other models,
as the agent positions are always constant here. It would be good to
support color coding them, but I'm not sure what a good way to integrate
that would be.