Add second Mason backend
This second backend is going to use a different strategy for double
buffering and execution of multiple step functions. I'm copying
the mason backend for now, as these are larger changes and I'd like
to have both for comparison.

Currently the new double buffering strategy is implemented, but the
step function scheduling is still as previously.