Avoid warnings due to circular targets
As of CMake 3.7 VERBOSE specifies the location the report is *copied*,
while the report location is given by REPORT_FILE. This is supposed
to be backwards compatible, but unfortunately causes make warnings if
the file passed to VERBOSE happens to be the default report file
location. Solve this by only passing the path to VERBOSE if we're using
CMake < 3.7.